moved harvested crop to data structures
