+++ RUFAS/routines/feed_storage/hay.py
@@ -1,10 +1,10 @@
 from RUFAS.current_day_conditions import CurrentDayConditions
+from RUFAS.data_structures.harvested_crop import HarvestedCrop
 from RUFAS.general_constants import GeneralConstants
 from RUFAS.time import Time
 from RUFAS.weather import Weather
 
 from .enums import CropCategory
-from .harvested_crop import HarvestedCrop
 from .storage import Storage
 
 """
