+++ RUFAS/routines/feed_storage/storage.py
@@ -1,6 +1,7 @@
 from typing import List
 
 from RUFAS.current_day_conditions import CurrentDayConditions
+from RUFAS.data_structures.harvested_crop import HarvestedCrop
 from RUFAS.general_constants import GeneralConstants
 from RUFAS.output_manager import OutputManager
 from RUFAS.time import Time
@@ -8,7 +9,6 @@ from RUFAS.units import MeasurementUnits
 from RUFAS.weather import Weather
 
 from .enums import CropCategory, CropType
-from .harvested_crop import HarvestedCrop
 
 """
 These constants define the upper and lower bounds of temperatures that allow fermentation (in degrees C), the upper and
